fix tests for new code flow
