+++ iridauploader/tests/core/test_upload.py
@@ -46,6 +46,14 @@ class TestUploadRunSingleEntry(unittest.TestCase):
         def message(self):
             return self._message
 
+        @property
+        def status(self):
+            return self._status
+        #
+        # @status.setter
+        # def status(self, status):
+        #     self._status = status
+
     def setUp(self):
         print("\nStarting " + self.__module__ + ": " + self._testMethodName)
         # config.setup()
@@ -297,12 +305,14 @@ class TestUploadRunSingleEntry(unittest.TestCase):
         # ensure upload occurs
         mock_validate_and_upload.assert_called_with(stub_directory_status, "mock_mode")
 
+    @patch("iridauploader.progress.upload_status._set_run_delayed")
     @patch("iridauploader.core.upload.upload_helpers")
     @patch("iridauploader.core.upload._validate_and_upload")
     @patch("iridauploader.core.upload.api_handler")
     @patch("iridauploader.core.upload.parsing_handler")
     def test_new_with_delay_config_directory_status(self, mock_parsing_handler, mock_api_handler,
-                                                    mock_validate_and_upload, mock_upload_helpers):
+                                                    mock_validate_and_upload, mock_upload_helpers,
+                                                    mock_set_run_delayed):
         """
         Checks that function exits with success when directory status is new and there is a delay set
         :return:
@@ -314,7 +324,7 @@ class TestUploadRunSingleEntry(unittest.TestCase):
         stub_directory_status._status = DirectoryStatus.NEW
 
         mock_parsing_handler.get_run_status.side_effect = [stub_directory_status]
-        mock_upload_helpers.set_run_delayed.side_effect = [None]
+        mock_set_run_delayed.side_effect = [None]
         mock_upload_helpers.directory_has_readonly_conflict.side_effect = [False]
 
         result = upload.upload_run_single_entry(stub_directory_status.directory)
@@ -324,11 +334,12 @@ class TestUploadRunSingleEntry(unittest.TestCase):
 
         # verify calls occurred
         mock_parsing_handler.get_run_status.assert_called_with(stub_directory_status.directory)
-        mock_upload_helpers.set_run_delayed.assert_called_with(stub_directory_status)
+        mock_set_run_delayed.assert_called_with(stub_directory_status)
 
         # ensure upload did not occur
         mock_validate_and_upload.assert_not_called()
 
+    # @patch("iridauploader.progress.upload_status._delayed_time_has_passed")
     @patch("iridauploader.core.upload.upload_helpers")
     @patch("iridauploader.core.upload._validate_and_upload")
     @patch("iridauploader.core.upload.api_handler")
@@ -361,12 +372,14 @@ class TestUploadRunSingleEntry(unittest.TestCase):
         # ensure upload occurs
         mock_validate_and_upload.assert_called_with(stub_directory_status, "mock_mode")
 
+    @patch("iridauploader.progress.upload_status._delayed_time_has_passed")
     @patch("iridauploader.core.upload.upload_helpers")
     @patch("iridauploader.core.upload._validate_and_upload")
     @patch("iridauploader.core.upload.api_handler")
     @patch("iridauploader.core.upload.parsing_handler")
     def test_delay_not_passed_directory_status(self, mock_parsing_handler, mock_api_handler,
-                                               mock_validate_and_upload, mock_upload_helpers):
+                                               mock_validate_and_upload, mock_upload_helpers,
+                                               mock_delayed_time_has_passed):
         """
         Checks that function exits with success when directory status is delayed and there is still delay time
         :return:
@@ -378,7 +391,7 @@ class TestUploadRunSingleEntry(unittest.TestCase):
         stub_directory_status._status = DirectoryStatus.DELAYED
 
         mock_parsing_handler.get_run_status.side_effect = [stub_directory_status]
-        mock_upload_helpers.delayed_time_has_passed.side_effect = [False]
+        mock_delayed_time_has_passed.side_effect = [False]
         mock_upload_helpers.directory_has_readonly_conflict.side_effect = [False]
 
         result = upload.upload_run_single_entry(stub_directory_status.directory)
@@ -388,7 +401,7 @@ class TestUploadRunSingleEntry(unittest.TestCase):
 
         # verify calls occurred
         mock_parsing_handler.get_run_status.assert_called_with(stub_directory_status.directory)
-        mock_upload_helpers.delayed_time_has_passed.assert_called_with(stub_directory_status, 1)
+        mock_delayed_time_has_passed.assert_called_with(stub_directory_status, 1)
 
         # ensure upload did not occur
         mock_validate_and_upload.assert_not_called()
