+++ iridauploader/tests/core/test_upload_helpers.py
@@ -1,8 +1,6 @@
 import unittest
 from unittest.mock import patch, MagicMock
 from os import path
-import time
-
 
 from iridauploader.api import UPLOAD_MODES, MODE_DEFAULT, MODE_FAST5, MODE_ASSEMBLIES
 from iridauploader.core import upload_helpers
@@ -95,62 +93,6 @@ class TestDirectoryHasReadonlyConflict(unittest.TestCase):
         self.assertEqual(False, upload_helpers.directory_has_readonly_conflict(""))
         self.assertEqual(False, upload_helpers.directory_has_readonly_conflict(""))
         self.assertEqual(True, upload_helpers.directory_has_readonly_conflict(""))
-
-
-class TestDelayedTimeHasPassed(unittest.TestCase):
-    """
-    Tests core.upload_helpers.delayed_time_has_passed
-    """
-
-    class StubDirectoryStatus:
-        time = None
-
-    def setUp(self):
-        print("\nStarting " + self.__module__ + ": " + self._testMethodName)
-
-    @patch("time.mktime")
-    def test_has_passed(self, mock_mktime):
-        """
-        Tests that True is returned when enough time has passed
-        :return:
-        """
-        # time.mktime will return the time float from 2 minutes ago
-        mock_mktime.side_effect = [time.time() - (2 * 60)]
-
-        stub_directory_status = self.StubDirectoryStatus()
-        stub_directory_status.time = time.strftime(DirectoryStatus.JSON_DATE_TIME_FORMAT)
-        # run was delayed for 1 minute
-        delay_minutes = 1
-
-        self.assertEqual(True, upload_helpers.delayed_time_has_passed(stub_directory_status, delay_minutes))
-
-    def test_delay_is_zero(self):
-        """
-        Tests that True is returned when there is no delay
-        :return:
-        """
-        stub_directory_status = self.StubDirectoryStatus()
-        stub_directory_status.time = time.strftime(DirectoryStatus.JSON_DATE_TIME_FORMAT)
-        # run was delayed for 1 minute
-        delay_minutes = 0
-
-        self.assertEqual(True, upload_helpers.delayed_time_has_passed(stub_directory_status, delay_minutes))
-
-    @patch("time.mktime")
-    def test_has_not_passed(self, mock_mktime):
-        """
-        Tests that False is returned when enough time has not passed
-        :return:
-        """
-        # time.mktime will return the time float from right now
-        mock_mktime.side_effect = [time.time()]
-
-        stub_directory_status = self.StubDirectoryStatus()
-        stub_directory_status.time = time.strftime(DirectoryStatus.JSON_DATE_TIME_FORMAT)
-        # run was delayed for 10 minute
-        delay_minutes = 10
-
-        self.assertEqual(False, upload_helpers.delayed_time_has_passed(stub_directory_status, delay_minutes))
 
 
 class TestParseAndValidate(unittest.TestCase):
